+++ app/src/main/java/vn/asiantech/way/data/source/remote/core/BaseServiceFactory.kt
@@ -1,0 +1,45 @@
+package vn.asiantech.way.data.source.remote.core
+
+import com.google.gson.FieldNamingPolicy
+import com.google.gson.Gson
+import com.google.gson.GsonBuilder
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+import vn.asiantech.way.BuildConfig
+import java.util.concurrent.TimeUnit
+
+/**
+ * Copyright © 2017 Asian Tech Co., Ltd.
+ * Created by quocnguyenp. on 11/30/17.
+ */
+open class BaseServiceFactory {
+
+    companion object {
+        private const val TIME_OUT = 120L
+    }
+
+    open fun makeLoggingInterceptor(): HttpLoggingInterceptor {
+        val logging = HttpLoggingInterceptor()
+        logging.level = if (BuildConfig.DEBUG)
+            HttpLoggingInterceptor.Level.BODY
+        else
+            HttpLoggingInterceptor.Level.NONE
+        return logging
+    }
+
+    open fun makeOkHttpClient(httpLoggingInterceptor: HttpLoggingInterceptor): OkHttpClient {
+        return OkHttpClient.Builder()
+                .addInterceptor(httpLoggingInterceptor)
+                .connectTimeout(TIME_OUT, TimeUnit.SECONDS)
+                .readTimeout(TIME_OUT, TimeUnit.SECONDS)
+                .build()
+    }
+
+    open fun makeGson(): Gson {
+        return GsonBuilder()
+                .setLenient()
+                .setDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'")
+                .setFieldNamingPolicy(FieldNamingPolicy.LOWER_CASE_WITH_UNDERSCORES)
+                .create()
+    }
+}
