+++ app/src/main/java/vn/asiantech/way/data/source/remote/googleapi/GoogleServiceFactory.kt
@@ -1,14 +1,11 @@
 package vn.asiantech.way.data.source.remote.googleapi
 
-import com.google.gson.FieldNamingPolicy
 import com.google.gson.Gson
-import com.google.gson.GsonBuilder
 import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
 import retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory
 import retrofit2.converter.gson.GsonConverterFactory
-import java.util.concurrent.TimeUnit
+import vn.asiantech.way.data.source.remote.core.BaseServiceFactory
 
 /**
  * Copyright © 2017 Asian Tech Co., Ltd.
@@ -17,14 +14,12 @@ import java.util.concurrent.TimeUnit
  * Provide "make" methods to create instances of [GoogleService]
  * and its related dependencies, such as OkHttpClient, Gson, etc.
  */
-object GoogleServiceFactory {
+object GoogleServiceFactory : BaseServiceFactory() {
 
     private const val BASE_URL = "https://maps.googleapis.com/maps/api/"
-    private const val TIME_OUT = 120L
 
-    fun makeGoogleService(isDebug: Boolean): GoogleService {
-        val okHttpClient = makeOkHttpClient(
-                makeLoggingInterceptor(isDebug))
+    fun getGoogleService(): GoogleService {
+        val okHttpClient = makeOkHttpClient(makeLoggingInterceptor())
         return makeGoogleService(okHttpClient, makeGson())
     }
 
@@ -36,30 +31,5 @@ object GoogleServiceFactory {
                 .addConverterFactory(GsonConverterFactory.create(gson))
                 .build()
         return retrofit.create(GoogleService::class.java)
-    }
-
-    fun makeGson(): Gson {
-        return GsonBuilder()
-                .setLenient()
-                .setDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'")
-                .setFieldNamingPolicy(FieldNamingPolicy.LOWER_CASE_WITH_UNDERSCORES)
-                .create()
-    }
-
-    fun makeLoggingInterceptor(isDebug: Boolean): HttpLoggingInterceptor {
-        val logging = HttpLoggingInterceptor()
-        logging.level = if (isDebug)
-            HttpLoggingInterceptor.Level.BODY
-        else
-            HttpLoggingInterceptor.Level.NONE
-        return logging
-    }
-
-    fun makeOkHttpClient(httpLoggingInterceptor: HttpLoggingInterceptor): OkHttpClient {
-        return OkHttpClient.Builder()
-                .addInterceptor(httpLoggingInterceptor)
-                .connectTimeout(TIME_OUT, TimeUnit.SECONDS)
-                .readTimeout(TIME_OUT, TimeUnit.SECONDS)
-                .build()
     }
 }
